modify destructor of rb and remove template
